docs: review sweep — remove dead de locale, guard docTags routes, extract HomePageBody, annotate search, fix a11y
1. Remove incomplete `de` locale block from i18n.ts (lines 166-207) and
   the `de: "de-DE"` entry from _doc-metainfo-area.tsx; `settings.locales`
   defines only `ja` so the `de` translations were unreachable and missing
   version.* and doc.pageCount* keys present in en/ja.

2. Tag routes (acceptable minimum): add a clear "feature incomplete — do not
   flip docTags until tag routes exist" guard comment co-located with both
   `enumerateTagsRoutes` in route-enumerators.ts and the `docTags` flag in
   settings.ts. No route modules added (large/risky; deferred).

3. Extract shared `HomePageBody({ locale, tree, categoryOrder, tagCount,
   overviewHref, tagsHref })` into pages/lib/_home-page-body.tsx and use it
   in both pages/index.tsx and pages/[locale]/index.tsx, eliminating ~95%
   duplication.

4. Search limitation (option b): add a comment in _search-widget-script.ts
   scoreEntry noting substring-only matching and deferred MiniSearch.

5. a11y: distinguish the "All Tags" h2 heading from the link in the docTags
   section — heading keeps t("doc.allTags") while the link now uses
   t("docs.browseAll") so the two interactive elements have distinct
   accessible names (WCAG 2.4.6).
