+++ docs/src/config/i18n.ts
@@ -163,48 +163,6 @@ const translations: Record<string, Record<string, string>> = {
     "version.page.status": "ステータス",
     "version.page.docs": "ドキュメント",
   },
-  de: {
-    "nav.gettingStarted": "Erste Schritte",
-    "nav.learn": "Lernen",
-    "nav.guides": "Anleitungen",
-    "nav.components": "Komponenten",
-    "nav.reference": "Referenz",
-    "nav.claude": "Claude",
-    "nav.changelog": "Changelog",
-    "nav.develop": "Entwicklung",
-    "nav.previous": "Zurück",
-    "nav.next": "Weiter",
-    "nav.overview": "Überblick",
-    "toc.title": "Auf dieser Seite",
-    "docs.browseAll": "Alle Dokumentationsabschnitte durchsuchen.",
-    "search.label": "Suche",
-    "search.resultCount": "{count} Ergebnisse",
-    "code.copy": "Code kopieren",
-    "code.copied": "Kopiert!",
-    "code.wrapToggle": "Zeilenumbruch umschalten",
-    "nav.backToMenu": "Zurück zum Hauptmenü",
-    "doc.editPage": "Diese Seite bearbeiten",
-    "doc.viewSource": "Quellcode auf GitHub ansehen",
-    "header.github": "GitHub-Repository",
-    "doc.tags": "Tags",
-    "doc.taggedWith": "Seiten mit Tag",
-    "doc.allTags": "Alle Tags",
-    "doc.created": "Erstellt",
-    "doc.updated": "Aktualisiert",
-    "doc.noTags": "Keine Tags gefunden.",
-    "doc.pageCount": "{count} Seiten",
-    "doc.pageCountSingle": "{count} Seite",
-    "doc.fallbackNotice":
-      "Diese Seite wurde noch nicht übersetzt und wird in der Originalsprache angezeigt.",
-    "frontmatter.preview.title": "Frontmatter",
-    "frontmatter.preview.keyCol": "Schlüssel",
-    "frontmatter.preview.valueCol": "Wert",
-    "version.latest": "Neueste",
-    "version.switcher.label": "Version",
-    "version.banner.unmaintained": "Sie sehen die Dokumentation einer älteren Version.",
-    "version.banner.unreleased": "Sie sehen unveröffentlichte Dokumentation.",
-    "version.banner.latestLink": "Neueste Version anzeigen",
-  },
 };
 
 /** Get a translated UI string */
